Move assessment classes to correct module
